[#963] separate middleware functions and add comment documentation
